This commit will include database support for adding bins.

+++ app/src/main/java/com/example/hackathonapp/ClothesBin.java
@@ -3,7 +3,7 @@ package com.example.hackathonapp;
 public class ClothesBin extends Bin{
 
 
-    public ClothesBin(int longitude, int latitude, String loc) {
+    public ClothesBin(double longitude, double latitude, String loc) {
         super(longitude, latitude, loc);
     }
 }
